#72 テスト用試合データ投入シードファイルを追加した
